// getway-api/src/campaign/models/Sms.model.ts
export interface SmsModel {
   
    msisdn: string
    incomingTime: string
    triggerId: string
    id: string
    triggerDescription: string
    isProcessed: number
    AFK_TRIGGERID_MSISDN: string
    trigger_attr_01?: string
    trigger_attr_02?: string
    trigger_attr_03?: string
    trigger_attr_04?: string
    trigger_attr_05?: string
    trigger_attr_06?: string
    trigger_attr_07?: string
    trigger_attr_08?: string
    trigger_attr_09?: string
    trigger_attr_10?: string
  
   




  }